# lib/strings: add parseLinuxConfig, linuxManualConfig: refactor readConfig