--- a/lib/strings.nix
+++ b/lib/strings.nix
@@ -35,6 +35,7 @@ rec {
     substring
     tail
     toJSON
+    tryEval
     typeOf
     unsafeDiscardStringContext
     ;
@@ -1148,6 +1149,68 @@ rec {
         absolutePaths = map (path: rootPath + "/${path}") relativePaths;
       in
         absolutePaths);
+
+  /*
+    Parse Linux Kernel Config text to an attribute set.
+    With `relaxed = true`, ill-formed lines will be skipped.
+
+    Example:
+      parseLinuxConfig {
+        configText = ''
+          # Kernel config
+          #
+
+          CONFIG_A=y
+          CONFIG_B=m
+          CONFIG_C="text"
+          CONFIG_D is not set
+        '';
+        relaxed = true;
+      }
+      => {
+        CONFIG_A = "y";
+        CONFIG_B = "m";
+        CONFIG_C = "text";
+      }
+  */
+  parseLinuxConfig = {
+    configText,
+    # skip ill-formed lines
+    relaxed ? false,
+  }:
+
+  let
+    inherit (lib) assertMsg nameValuePair listToAttrs;
+    rawLines = splitString "\n" configText;
+    nonEmptyLines = filter (line: line != "") rawLines;
+    configLines = filter (line: !(hasPrefix "#" line)) nonEmptyLines;
+    parseConfigLine =
+      line:
+      let
+        parsed = parseConfigLineUnquoted line;
+      in
+      parsed // { value = removeQuote parsed.value; };
+    parseConfigLineUnquoted =
+      line:
+      let
+        parsed = match "(CONFIG_[^=]+)=(.*)$" line;
+      in
+      assert assertMsg (parsed != null) "failed to parse kernel config line '${line}'";
+      nameValuePair (elemAt parsed 0) (elemAt parsed 1);
+    removeQuote =
+      value:
+      let
+        quoteMatch = match ''"(.*)"'' value;
+      in
+      if quoteMatch == null then value else elemAt quoteMatch 0;
+    tryParseConfigLine = line: tryEval (parseConfigLine line);
+  in
+  if relaxed then
+    listToAttrs (
+      map (result: result.value) (filter (result: result.success) (map tryParseConfigLine configLines))
+    )
+  else
+    listToAttrs (map parseConfigLine configLines);
 
   /* Read the contents of a file removing the trailing \n