@shlevy 2024-03-29
Would be curious to see some profiling results of this vs the IFD approach on a real config.