@lolbinarycat 2024-03-30
one advantage of IFD is it actually allows expensive computations to be cached.  not sure if parsing a trivial format counts as "expensive computation", but it's something to keep in mind when benchmarking.