@linyinfeng 2024-03-30
> Would be curious to see some profiling results of this vs the IFD approach on a real config.

> one advantage of IFD is it actually allows expensive computations to be cached. not sure if parsing a trivial format counts as "expensive computation", but it's something to keep in mind when benchmarking.

I have done some benchmarks with `hyperfine`. Performance with `parseLinuxConfig` is almost the same as the performance of the IFD approach when `config.nix` has been cached (on my laptop with the kernel config of my laptop).

```
Benchmark 1: nix eval .#lib.configOriginal --option eval-cache false >/dev/null
  Time (mean ± σ):     330.8 ms ±  38.7 ms    [User: 218.0 ms, System: 61.3 ms]
  Range (min … max):   303.1 ms … 435.4 ms    10 runs
 
Benchmark 2: nix eval .#lib.configNew --option eval-cache false >/dev/null
  Time (mean ± σ):     314.2 ms ±   8.1 ms    [User: 230.9 ms, System: 57.0 ms]
  Range (min … max):   304.7 ms … 328.6 ms    10 runs
```

Here is the benchmark script:
 <https://gist.github.com/linyinfeng/099b6061a08fc381e28c7a3eaa612f01>.

```nix
{
  inputs = { nixpkgs-299926.url = "github:nixos/nixpkgs/pull/299926/head"; };
  outputs = { self, nixpkgs-299926 }:
    let
      testSystem = "x86_64-linux";
      pkgs = nixpkgs-299926.legacyPackages.${testSystem};
      inherit (pkgs) runCommand lib writeShellApplication;
      inherit (builtins) toJSON;
      # from https://github.com/NixOS/nixpkgs/blob/master/pkgs/os-specific/linux/kernel/manual-config.nix
      readConfigOriginal = configfile:
        import (runCommand "config.nix" { } ''
          echo "{" > "$out"
          while IFS='=' read key val; do
            [ "x''${key#CONFIG_}" != "x$key" ] || continue
            no_firstquote="''${val#\"}";
            echo '  "'"$key"'" = "'"''${no_firstquote%\"}"'";' >> "$out"
          done < "${configfile}"
          echo "}" >> $out
        '').outPath;
      # from https://github.com/NixOS/nixpkgs/pull/299926/files
      readConfigNew = configfile:
        lib.strings.parseLinuxConfig {
          # IFD could happen in `builtins.readFile`
          configText = builtins.readFile configfile;
          relaxed = true;
        };
    in {
      lib = {
        configOriginal = readConfigOriginal ./example.config;
        configNew = readConfigNew ./example.config;
      };
      packages.${testSystem} = {
        config-diff = runCommand "config-diff" { } ''
          diff <(${toJSON self.lib.configOriginal}) <(${
            toJSON self.lib.configNew
          }) > $out
        '';
        test = writeShellApplication {
          name = "test";
          runtimeInputs = with pkgs; [ hyperfine ];
          text = let
            # disable flake eval cache
            cmd = item:
              "nix eval .#lib.${item} --option eval-cache false >/dev/null";
            cmdEmpty =
              "nix eval --expr '{}' --option eval-cache false >/dev/null";
          in ''
            echo -n "lines: "
            wc -l example.config
            echo "-- original: first time eval --"
            time ${cmd "configOriginal"}
            echo "-- new:      first time eval --"
            time ${cmd "configNew"}
            echo "-- hyperfine (--warmup 3) --"
            hyperfine --warmup 3 "${cmd "configOriginal"}" "${
              cmd "configNew"
            }" "${cmdEmpty}"
          '';
        };
      };
    };
}
```

Note that I passed `--option eval-cache false` to disable flake eval cache.

Here are the full benchmark console logs. I benchmarked twice. In the first benchmark `configNew` is slightly faster than `configOriginal`, and in the second benchmark `configOriginal` is slightly faster.

```console
yinfeng@framework ~/S/p/clean> nix --version
nix (Nix) 2.18.2
yinfeng@framework ~/S/p/clean> curl https://gist.githubusercontent.com/linyinfeng/099b6061a08fc381e28c7a3eaa612f01/raw/8f60ab7dad1403c9066282c6660d391d440c8725/flake.nix > flake.nix
  % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                 Dload  Upload   Total   Spent    Left  Speed
100  2313  100  2313    0     0   4427      0 --:--:-- --:--:-- --:--:--  4431
yinfeng@framework ~/S/p/clean> nix flake lock
warning: creating lock file '/home/yinfeng/Source/playground-benchmark-parse-linux-config/clean/flake.lock'
yinfeng@framework ~/S/p/clean> cat /proc/config.gz | gzip -d > example.config
yinfeng@framework ~/S/p/clean> head example.config
#
# Automatically generated file; DO NOT EDIT.
# Linux/x86_64 6.6.14 Kernel Configuration
#
CONFIG_CC_VERSION_TEXT="gcc (GCC) 13.2.0"
CONFIG_CC_IS_GCC=y
CONFIG_GCC_VERSION=130200
CONFIG_CLANG_VERSION=0
CONFIG_AS_IS_GNU=y
CONFIG_AS_VERSION=24100
yinfeng@framework ~/S/p/clean> echo "# some random comment, so that we can test the first time performance" >> example.config 
yinfeng@framework ~/S/p/clean> tail example.config 
CONFIG_MEMTEST=y
# CONFIG_HYPERV_TESTING is not set
# end of Kernel Testing and Coverage

#
# Rust hacking
#
# end of Rust hacking
# end of Kernel hacking
# some random comment, so that we can test the first time performance
yinfeng@framework ~/S/p/clean> wc -l example.config
11712 example.config
yinfeng@framework ~/S/p/clean> nix run .#test
lines: 11712 example.config
-- original: first time eval --

real	0m6.105s
user	0m0.221s
sys	0m0.084s
-- new:      first time eval --

real	0m0.332s
user	0m0.231s
sys	0m0.074s
-- hyperfine (--warmup 3) --
Benchmark 1: nix eval .#lib.configOriginal --option eval-cache false >/dev/null
  Time (mean ± σ):     330.8 ms ±  38.7 ms    [User: 218.0 ms, System: 61.3 ms]
  Range (min … max):   303.1 ms … 435.4 ms    10 runs
 
Benchmark 2: nix eval .#lib.configNew --option eval-cache false >/dev/null
  Time (mean ± σ):     314.2 ms ±   8.1 ms    [User: 230.9 ms, System: 57.0 ms]
  Range (min … max):   304.7 ms … 328.6 ms    10 runs
 
Benchmark 3: nix eval --expr '{}' --option eval-cache false >/dev/null
  Time (mean ± σ):      16.1 ms ±   1.5 ms    [User: 9.7 ms, System: 5.9 ms]
  Range (min … max):    13.1 ms …  20.9 ms    170 runs
 
Summary
  nix eval --expr '{}' --option eval-cache false >/dev/null ran
   19.53 ± 1.94 times faster than nix eval .#lib.configNew --option eval-cache false >/dev/null
   20.57 ± 3.12 times faster than nix eval .#lib.configOriginal --option eval-cache false >/dev/null
yinfeng@framework ~/S/p/clean> nix build .#config-diff
yinfeng@framework ~/S/p/clean> cat ./result
yinfeng@framework ~/S/p/clean> # diff is empty!
yinfeng@framework ~/S/p/clean> nix run .#test
lines: 11712 example.config
-- original: first time eval --

real	0m0.322s
user	0m0.215s
sys	0m0.064s
-- new:      first time eval --

real	0m0.535s
user	0m0.280s
sys	0m0.060s
-- hyperfine (--warmup 3) --
Benchmark 1: nix eval .#lib.configOriginal --option eval-cache false >/dev/null
  Time (mean ± σ):     315.7 ms ±   8.0 ms    [User: 206.3 ms, System: 68.0 ms]
  Range (min … max):   303.7 ms … 326.1 ms    10 runs
 
Benchmark 2: nix eval .#lib.configNew --option eval-cache false >/dev/null
  Time (mean ± σ):     316.4 ms ±  14.6 ms    [User: 230.3 ms, System: 61.0 ms]
  Range (min … max):   304.0 ms … 352.6 ms    10 runs
 
Benchmark 3: nix eval --expr '{}' --option eval-cache false >/dev/null
  Time (mean ± σ):      15.9 ms ±   1.5 ms    [User: 10.0 ms, System: 5.5 ms]
  Range (min … max):    13.1 ms …  20.5 ms    178 runs
 
Summary
  nix eval --expr '{}' --option eval-cache false >/dev/null ran
   19.83 ± 1.99 times faster than nix eval .#lib.configOriginal --option eval-cache false >/dev/null
   19.88 ± 2.14 times faster than nix eval .#lib.configNew --option eval-cache false >/dev/null
```